requirements.txt
# Core Data Processing
pandas==2.2.0
pyarrow==15.0.0
numpy==1.26.3

# Google Cloud Platform
google-cloud-storage==2.14.0
google-cloud-bigquery==3.17.2

# dbt
dbt-core==1.7.8
dbt-bigquery==1.7.4

# Testing
pytest==8.0.0
pytest-cov==4.1.0

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Utilities
python-dotenv==1.0.1
